location post and delete handlers have been moved to separate modules
